--- src/main/java/org/sesac/slopedbe/auth/service/KakaoLoginService.java
@@ -1,66 +1,0 @@
-package org.sesac.slopedbe.auth.service;
-
-import org.sesac.slopedbe.auth.model.dto.response.KakaoTokenResponse;
-import org.sesac.slopedbe.auth.model.dto.response.KakaoUserInfoResponse;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatusCode;
-import org.springframework.stereotype.Service;
-import org.springframework.web.reactive.function.client.WebClient;
-
-import io.netty.handler.codec.http.HttpHeaderValues;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import reactor.core.publisher.Mono;
-
-@Slf4j
-@RequiredArgsConstructor
-@Service
-public class KakaoLoginService {
-
-	@Value("${spring.security.oauth2.client.registration.kakao.client-id}")
-	private String clientId;
-	@Value("${spring.security.oauth2.client.registration.kakao.client-secret}")
-	private String clientSecret;
-
-	private final String KAUTH_TOKEN_URL_HOST = "https://kauth.kakao.com";
-	private final String KAUTH_USER_URL_HOST = "https://kapi.kakao.com";
-
-	public String getAccessTokenFromKakao(String code) {
-		KakaoTokenResponse kakaoTokenResponse = WebClient.create(KAUTH_TOKEN_URL_HOST).post()
-			.uri(uriBuilder -> uriBuilder
-				.scheme("https")
-				.path("/oauth/token")
-				.queryParam("grant_type", "authorization_code")
-				.queryParam("client_id", clientId)
-				.queryParam("code", code)
-				.queryParam("client_secret",clientSecret)
-				.build(true))
-			.header(HttpHeaders.CONTENT_TYPE, HttpHeaderValues.APPLICATION_X_WWW_FORM_URLENCODED.toString())
-			.retrieve()
-			.onStatus(HttpStatusCode::is4xxClientError, clientResponse -> Mono.error(new RuntimeException("Invalid Parameter")))
-			.onStatus(HttpStatusCode::is5xxServerError, clientResponse -> Mono.error(new RuntimeException("Internal Server Error")))
-			.bodyToMono(KakaoTokenResponse.class)
-			.block();
-
-		return kakaoTokenResponse.getAccessToken();
-	}
-
-	public KakaoUserInfoResponse getUserInfo(String accessToken) {
-		KakaoUserInfoResponse userInfo = WebClient.create(KAUTH_USER_URL_HOST)
-			.get()
-			.uri(uriBuilder -> uriBuilder
-				.scheme("https")
-				.path("/v2/user/me")
-				.build(true))
-			.header(HttpHeaders.AUTHORIZATION, "Bearer " + accessToken)
-			.header(HttpHeaders.CONTENT_TYPE, HttpHeaderValues.APPLICATION_X_WWW_FORM_URLENCODED.toString())
-			.retrieve()
-			.onStatus(HttpStatusCode::is4xxClientError, clientResponse -> Mono.error(new RuntimeException("Invalid Parameter")))
-			.onStatus(HttpStatusCode::is5xxServerError, clientResponse -> Mono.error(new RuntimeException("Internal Server Error")))
-			.bodyToMono(KakaoUserInfoResponse.class)
-			.block();
-
-		return userInfo;
-	}
-}
